// src/modules/webtoon/webtoon/models/enums/image-types.ts
/* eslint-disable @/no-unused-vars */
enum ImageTypes{
    WEBTOON_THUMBNAIL = "WEBTOON_THUMBNAIL",
    WEBTOON_BACKGROUND_BANNER = "WEBTOON_BACKGROUND_BANNER",
    WEBTOON_TOP_BANNER = "WEBTOON_TOP_BANNER",
    WEBTOON_MOBILE_BANNER = "WEBTOON_MOBILE_BANNER",
    EPISODE_THUMBNAIL = "EPISODE_THUMBNAIL",
    EPISODE_IMAGE = "EPISODE_IMAGE",
}
export default ImageTypes;